Split audit package data helpers
Move `pnpm ls` package-tree loading and bulk advisory normalisation
out of `audit-utils.js` into a focused sibling module. Keep
`audit-utils.js` as the audit orchestration layer and re-export the
helper APIs for existing callers.
