+++ security/audit-utils.js
@@ -14,19 +14,29 @@
  */
 
 import { execFileSync, spawnSync } from 'node:child_process';
+import {
+  collectInstalledPackageVersions,
+  normalizeBulkAdvisories,
+  parseJsonOutput,
+} from './audit-package-data.js';
 
+export {
+  buildVersionMap,
+  collectInstalledPackageVersions,
+  loadPackageTrees,
+  normalizeBulkAdvisories,
+  parseJsonOutput,
+} from './audit-package-data.js';
 export {
   partitionAdvisoriesById,
   reportUnexpectedAdvisories,
 } from './audit-reporting.js';
 
 const AUDIT_ARGS = ['audit', '--json'];
-const LIST_ARGS = ['ls', '--json', '--depth', 'Infinity'];
 const BULK_ADVISORY_PATH = '-/npm/v1/security/advisories/bulk';
 const BULK_AUDIT_TIMEOUT_MS = 30_000;
 const DEFAULT_REGISTRY = 'https://registry.npmjs.org/';
 const COMMAND_MAX_BUFFER = 64 * 1024 * 1024;
-const DEPENDENCY_SECTION_NAMES = ['dependencies', 'devDependencies', 'optionalDependencies'];
 const RETIRED_AUDIT_ENDPOINT_MESSAGE = 'This endpoint is being retired. Use the bulk advisory endpoint instead.';
 const defaultAuditIo = {
   execFileSync,
@@ -35,27 +45,6 @@ const defaultAuditIo = {
   clearTimeout,
   spawnSync,
 };
-
-/** Parse command JSON and optionally reject blank responses.
- * @param {string | undefined | null} payloadText Raw command output. @param {string} commandLabel Label used in parse errors. @param {{ requireNonEmpty?: boolean }} [options={}] Parsing options.
- * @returns {Record<string, unknown> | unknown[]} Parsed JSON value, or `{}` for optional blank output. @example parseJsonOutput('{"advisories":{}}', 'pnpm audit'); // { advisories: {} }
- */
-function parseJsonOutput(payloadText, commandLabel, options = {}) {
-  const { requireNonEmpty = false } = options;
-  const text = payloadText?.trim?.() ?? '';
-  if (!text) {
-    if (requireNonEmpty) {
-      throw new Error(`Failed to parse ${commandLabel} JSON: response body was empty.`);
-    }
-    return {};
-  }
-  try {
-    return JSON.parse(text);
-  } catch (error) {
-    error.message = `Failed to parse ${commandLabel} JSON: ${error.message}`;
-    throw error;
-  }
-}
 
 /** Detect whether pnpm reported the retired audit endpoint.
  * @param {unknown} payload Parsed `pnpm audit --json` payload.
@@ -66,86 +55,6 @@ function isRetiredAuditEndpoint(payload) {
     payload?.error?.code === 'ERR_PNPM_AUDIT_BAD_RESPONSE' &&
     typeof payload?.error?.message === 'string' &&
     payload.error.message.includes(RETIRED_AUDIT_ENDPOINT_MESSAGE));
-}
-
-/** Check whether a version points at a local workspace dependency.
- * @param {string} version Package version or workspace reference.
- * @returns {boolean} `true` when the version should be ignored for registry audits. @example isLocalWorkspaceVersion('workspace:*'); // true
- */
-function isLocalWorkspaceVersion(version) {
-  return (
-    version.startsWith('file:') ||
-    version.startsWith('link:') ||
-    version.startsWith('workspace:'));
-}
-
-/** Record an installed package version unless it is missing or workspace-local.
- * @param {Map<string, Set<string>>} versionsByPackage Installed versions keyed by package name. @param {string} packageName Package name from `pnpm ls`. @param {string} version Installed package version.
- * @returns {void} @example const versions = new Map(); addPackageVersion(versions, 'validator', '13.15.23'); console.log([...versions.get('validator')]); // ['13.15.23']
- */
-function addPackageVersion(versionsByPackage, packageName, version) {
-  const isMissing = !packageName || !version;
-  if (isMissing || isLocalWorkspaceVersion(version)) {
-    return;
-  }
-  const knownVersions = versionsByPackage.get(packageName) ?? new Set();
-  knownVersions.add(version);
-  versionsByPackage.set(packageName, knownVersions);
-}
-
-/** Walk one dependency section from `pnpm ls` and record installed versions.
- * @param {Record<string, unknown> | undefined} section Dependency section keyed by package name. @param {Map<string, Set<string>>} versionsByPackage Collected versions keyed by package name.
- * @returns {void} @example const versions = new Map(); walkDependencySection({ validator: { version: '13.15.23' } }, versions); console.log(versions.get('validator').has('13.15.23')); // true
- */
-function walkDependencySection(section, versionsByPackage) {
-  if (!section || typeof section !== 'object') {
-    return;
-  }
-  for (const [packageName, dependency] of Object.entries(section)) {
-    if (!dependency || typeof dependency !== 'object') {
-      continue;
-    }
-    if (typeof dependency.version === 'string') {
-      addPackageVersion(versionsByPackage, packageName, dependency.version);
-    }
-    walkDependencies(dependency, versionsByPackage);
-  }
-}
-
-/** Walk a `pnpm ls` tree and record every installed package version.
- * @param {Record<string, unknown> | undefined} node Dependency tree node from `pnpm ls`. @param {Map<string, Set<string>>} versionsByPackage Collected versions keyed by package name.
- * @returns {void} @example const versions = new Map(); walkDependencies({ dependencies: { validator: { version: '13.15.23' } } }, versions); console.log([...versions.get('validator')]); // ['13.15.23']
- */
-function walkDependencies(node, versionsByPackage) {
-  if (!node || typeof node !== 'object') {
-    return;
-  }
-  for (const sectionName of DEPENDENCY_SECTION_NAMES) {
-    walkDependencySection(node[sectionName], versionsByPackage);
-  }
-}
-
-/** Check whether `pnpm ls` returned a dependency tree object.
- * @param {unknown} value Parsed `pnpm ls` payload value.
- * @returns {boolean} `true` when the value can be walked as one dependency tree. @example isDependencyTreeNode({ dependencies: {} }); // true
- */
-function isDependencyTreeNode(value) {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-/** Build the installed package-version map from parsed `pnpm ls` output.
- * @param {Record<string, unknown> | Record<string, unknown>[] | undefined} packageTrees Parsed `pnpm ls` output as one tree or many. @returns {Map<string, Set<string>>} Installed versions keyed by package name. @example const versions = buildVersionMap([{ dependencies: { validator: { version: '13.15.23' } } }]); console.log(versions.get('validator').has('13.15.23')); // true
- */
-function buildVersionMap(packageTrees) {
-  const versionsByPackage = new Map();
-  const trees = Array.isArray(packageTrees) ? packageTrees : [packageTrees];
-  for (const tree of trees) {
-    if (!isDependencyTreeNode(tree)) {
-      throw new TypeError('pnpm ls returned an invalid dependency tree payload.');
-    }
-    walkDependencies(tree, versionsByPackage);
-  }
-  return versionsByPackage;
 }
 
 /** Ensure a child-process result exited normally.
@@ -165,36 +74,6 @@ function assertCompletedProcess(result, commandLabel) {
     throw new Error(`${commandLabel} was terminated before reporting an exit status.`);
   }
   return result.status;
-}
-
-/** Collect installed package versions from `pnpm ls` for bulk advisory lookups.
- * Throws when `pnpm ls` fails, is signalled, or returns invalid JSON.
- * @param {object} [auditIo=defaultAuditIo] Audit IO adapter; `defaultAuditIo` is used when omitted.
- * @returns {Record<string, string[]>} Sorted installed versions keyed by package name.
- * @example // Normal usage with `pnpm ls` returning one installed validator version:
- * collectInstalledPackageVersions(); // { validator: ['13.15.23'] }
- * @example const auditIo = { ...defaultAuditIo, spawnSync: () => ({ status: 0, stdout: '[{"dependencies":{"validator":{"version":"13.15.23"}}}]' }) };
- * collectInstalledPackageVersions(auditIo); // { validator: ['13.15.23'] }
- */
-function collectInstalledPackageVersions(auditIo = defaultAuditIo) {
-  const result = auditIo.spawnSync('pnpm', LIST_ARGS, { encoding: 'utf8', maxBuffer: COMMAND_MAX_BUFFER, stdio: ['ignore', 'pipe', 'inherit'] });
-  const status = assertCompletedProcess(result, 'pnpm ls');
-  if (status !== 0) {
-    throw new Error(`pnpm ls failed without producing a dependency tree (exit status ${status}).`);
-  }
-  const stdout = result.stdout?.trim();
-  if (!stdout) {
-    throw new Error('pnpm ls failed without producing a dependency tree.');
-  }
-
-  const packageTrees = parseJsonOutput(stdout, 'pnpm ls');
-  const versionsByPackage = buildVersionMap(packageTrees);
-
-  return Object.fromEntries(
-    [...versionsByPackage.entries()]
-      .sort(([left], [right]) => left.localeCompare(right))
-      .map(([packageName, versions]) => [packageName, [...versions].sort()]),
-  );
 }
 
 /** Return `true` when a raw registry string is a real URL and not a placeholder.
@@ -233,78 +112,6 @@ function readRegistryUrl(auditIo = defaultAuditIo) {
   } catch {
     return DEFAULT_REGISTRY;
   }
-}
-
-/** Extract a GitHub advisory identifier from an advisory URL.
- * @param {unknown} advisoryUrl Advisory URL from pnpm or npm audit output.
- * @returns {string | undefined} Matching GHSA identifier when one is present. @example extractGithubAdvisoryId('https://github.com/advisories/GHSA-vghf-hv5q-vc2g'); // 'GHSA-vghf-hv5q-vc2g'
- */
-function extractGithubAdvisoryId(advisoryUrl) {
-  if (typeof advisoryUrl !== 'string') {
-    return undefined;
-  }
-  const match = advisoryUrl.match(/GHSA-([0-9a-z]{4})-([0-9a-z]{4})-([0-9a-z]{4})/i);
-  if (!match) {
-    return undefined;
-  }
-  const [, first, second, third] = match;
-  return `GHSA-${first.toLowerCase()}-${second.toLowerCase()}-${third.toLowerCase()}`;
-}
-
-/** Derive the advisory key used to deduplicate bulk advisory responses.
- * @param {string} packageName Advisory package name. @param {{ id?: unknown, url?: unknown }} advisory Raw advisory object.
- * @returns {{ key: string, githubAdvisoryId: string | undefined }} Stable advisory key and extracted GHSA identifier. @example deriveAdvisoryKey('validator', { id: 100000, url: 'https://github.com/advisories/GHSA-vghf-hv5q-vc2g' }); // { key: 'GHSA-vghf-hv5q-vc2g', githubAdvisoryId: 'GHSA-vghf-hv5q-vc2g' }
- */
-function deriveAdvisoryKey(packageName, advisory) {
-  const githubAdvisoryId = extractGithubAdvisoryId(advisory?.url);
-  const key = githubAdvisoryId ?? `${packageName}:${String(advisory?.id ?? 'unknown')}`;
-  return { key, githubAdvisoryId };
-}
-
-/** Return `true` when a value is a plain (non-array, non-null) object.
- * @param {unknown} value Value to test. @returns {boolean}
- * @example isPlainAdvisoryObject({ id: 1 }); // true
- */
-function isPlainAdvisoryObject(value) { return typeof value === 'object' && value !== null && !Array.isArray(value); }
-
-/** Merge advisories for one package into the shared accumulator.
- * @param {string} packageName Package name from the bulk advisory payload. @param {unknown[]} packageAdvisories Validated array of raw advisory objects. @param {Record<string, unknown>} advisories Accumulator mutated in place.
- * @returns {void} @example const advisories = {}; addPackageAdvisories('validator', [{ id: 100000, url: 'https://github.com/advisories/GHSA-vghf-hv5q-vc2g', title: 'Validator SSRF' }], advisories); console.log(advisories['GHSA-vghf-hv5q-vc2g'].package_name); // 'validator'
- */
-function addPackageAdvisories(packageName, packageAdvisories, advisories) {
-  for (const [index, advisory] of packageAdvisories.entries()) {
-    if (!isPlainAdvisoryObject(advisory)) {
-      throw new Error(`Invalid advisory for package ${packageName} at index ${index}: expected object`);
-    }
-    const { key, githubAdvisoryId } = deriveAdvisoryKey(packageName, advisory);
-    if (Object.hasOwn(advisories, key)) {
-      continue;
-    }
-    advisories[key] = {
-      ...advisory,
-      github_advisory_id: githubAdvisoryId,
-      package_name: packageName,
-    };
-  }
-}
-
-/** Normalize bulk advisory responses into the shared advisory object shape.
- * @param {Record<string, unknown> | undefined} bulkPayload Bulk advisory payload keyed by package name.
- * @returns {Record<string, unknown>} Deduplicated advisories keyed by GHSA identifier or package fallback. @example normalizeBulkAdvisories({ validator: [{ id: 100000, url: 'https://github.com/advisories/GHSA-vghf-hv5q-vc2g' }] }); // { 'GHSA-vghf-hv5q-vc2g': { github_advisory_id: 'GHSA-vghf-hv5q-vc2g', package_name: 'validator', id: 100000, url: 'https://github.com/advisories/GHSA-vghf-hv5q-vc2g' } }
- */
-function normalizeBulkAdvisories(bulkPayload) {
-  if (!isPlainAdvisoryObject(bulkPayload)) {
-    throw new TypeError('Invalid bulk advisory payload: expected an object keyed by package name.');
-  }
-  const advisories = {};
-  for (const [packageName, packageAdvisories] of Object.entries(bulkPayload)) {
-    if (!Array.isArray(packageAdvisories)) {
-      throw new TypeError(`Invalid bulk advisory entry for package ${packageName}: expected array, received ${JSON.stringify(packageAdvisories)}`);
-    }
-    addPackageAdvisories(packageName, packageAdvisories, advisories);
-  }
-
-  return advisories;
 }
 
 /** Post package versions to the npm bulk advisory endpoint and return the raw response.
@@ -359,7 +166,7 @@ async function runBulkAdvisoryAudit(auditIo = defaultAuditIo) {
   const endpoint = new URL(BULK_ADVISORY_PATH, registryUrl);
   const { response, responseText } = await fetchBulkAdvisories(
     endpoint,
-    collectInstalledPackageVersions(auditIo),
+    collectInstalledPackageVersions(auditIo, assertCompletedProcess),
     auditIo,
   );
